refactor seller details, read seller orders
