+++ pages/seller_sign_up.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import Header from "../components/Common/Header";
 import ShopNav from "../components/layouts/ShopNav";
 import ProtectedRoute from "../components/route/ProtectedRoute";
-import SellerSignUpForm from "../components/Seller/SellerSignUpForm";
+import SellerDetails, { SIGN_UP_TYPE } from "../components/Seller/SellerDetails";
 
 const SellerSignUp = () => {
   return (
@@ -10,7 +10,7 @@ const SellerSignUp = () => {
       <ShopNav>
         <Header name="Seller Sign Up Form" />
         <p className="ml-6 mt-4">Fill in some business details and verify your identity with Stripe to receive payouts for your completed sales.</p>
-        <SellerSignUpForm />
+        <SellerDetails callerType={SIGN_UP_TYPE}/>
       </ShopNav>
     </ProtectedRoute>
   );
